change game logic to single file
